chore(test): fold vitest.shared into the root config
