+++ AGENTS.md
@@ -30,7 +30,7 @@ Per-package: `tsc --noEmit && vitest run --coverage`. A root `vitest` run covers
 
 ## Testing
 
-Runner: **vitest** with istanbul coverage. Test config is centralized in root `vitest.shared.ts` - per-package `vitest.config.ts` files are two-liners calling `suite()` (or `suite(true)` for DOM packages, which run in vitest's happy-dom environment). Coverage: every package gates statements, branches, functions, and lines at 100% via the shared `coverage.thresholds` - a conditional arm is either exercised by a test or deleted as dead code. The `PR` workflow blocks merges on these thresholds; the README badge is generated by `.github/workflows/coverage.yml` from measured lcov on pushes to `main`.
+Runner: **vitest** with istanbul coverage. Test config is centralized in root `vitest.config.ts`, which defines the workspace projects and exports `suite()` - per-package `vitest.config.ts` files are two-liners calling `suite()` (or `suite(true)` for DOM packages, which run in vitest's happy-dom environment). Coverage: every package gates statements, branches, functions, and lines at 100% via the shared `coverage.thresholds` - a conditional arm is either exercised by a test or deleted as dead code. The `PR` workflow blocks merges on these thresholds; the README badge is generated by `.github/workflows/coverage.yml` from measured lcov on pushes to `main`.
 
 ```ts
 // Test primitives come directly from vitest
