+++ packages/router/vitest.config.ts
@@ -1,3 +1,3 @@
-import { suite } from '../../vitest.shared';
+import { suite } from '../../vitest.config';
 
 export default suite(true);
